[MIG] company_dependent_flag: Migration to 18.0
