--- company_dependent_flag/__init__.py
@@ -1,6 +1,0 @@
-from odoo.tools import config
-
-from . import models
-
-if not config.get("without_demo"):
-    from . import demo
