--- company_dependent_flag/demo/__init__.py
@@ -1,1 +1,0 @@
-from . import partner
