--- company_dependent_flag/demo/partner.py
@@ -1,9 +1,0 @@
-# © 2023 David BEAL @ Akretion
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from odoo import fields, models
-
-
-class ResPartner(models.Model):
-    _inherit = "res.partner"
-
-    phone = fields.Char(company_dependent=True)
