--- company_dependent_flag/readme/CONTRIBUTORS.md
@@ -1,2 +1,0 @@
-- [Akretion](https://akretion.com):  
-  - David BEAL \<david.beal@akretion.com\>
