--- company_dependent_flag/readme/ROADMAP.md
@@ -1,2 +1,0 @@
-- In the next versions, the module should be renamed
-  web_field_company_dependent_flag
