+++ web_field_company_dependent_flag/README.rst
@@ -17,10 +17,10 @@ Company Dependent Flag
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fmulti--company-lightgray.png?logo=github
-    :target: https://github.com/OCA/multi-company/tree/18.0/company_dependent_flag
+    :target: https://github.com/OCA/multi-company/tree/18.0/web_field_company_dependent_flag
     :alt: OCA/multi-company
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/multi-company-18-0/multi-company-18-0-company_dependent_flag
+    :target: https://translation.odoo-community.org/projects/multi-company-18-0/multi-company-18-0-web_field_company_dependent_flag
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/builds?repo=OCA/multi-company&target_branch=18.0
@@ -33,18 +33,12 @@ adhoc class.
 
 |image1|
 
-.. |image1| image:: https://raw.githubusercontent.com/OCA/multi-company/18.0/company_dependent_flag/static/description/demo.png
+.. |image1| image:: https://raw.githubusercontent.com/OCA/multi-company/18.0/web_field_company_dependent_flag/static/description/demo.png
 
 **Table of contents**
 
 .. contents::
    :local:
-
-Known issues / Roadmap
-======================
-
-- In the next versions, the module should be renamed
-  web_field_company_dependent_flag
 
 Bug Tracker
 ===========
@@ -52,7 +46,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/multi-company/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/multi-company/issues/new?body=module:%20company_dependent_flag%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/multi-company/issues/new?body=module:%20web_field_company_dependent_flag%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -71,6 +65,10 @@ Contributors
 
   - David BEAL <david.beal@akretion.com>
 
+- `PyTech <https://www.pytech.it>`__:
+
+  - Simone Rubino <simone.rubino@pytech.it>
+
 Maintainers
 -----------
 
@@ -84,6 +82,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/multi-company <https://github.com/OCA/multi-company/tree/18.0/company_dependent_flag>`_ project on GitHub.
+This module is part of the `OCA/multi-company <https://github.com/OCA/multi-company/tree/18.0/web_field_company_dependent_flag>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
