+++ web_field_company_dependent_flag/__init__.py
@@ -1,0 +1,1 @@
+from . import models
