+++ web_field_company_dependent_flag/__manifest__.py
@@ -1,8 +1,8 @@
 {
     "name": "Company Dependent Flag",
-    "author": "Akretion,Odoo Community Association (OCA)",
-    "summary": "Apply css style to company dependent fields",
-    "version": "16.0.1.0.2",
+    "author": "Akretion, Odoo Community Association (OCA)",
+    "summary": "Apply CSS style to company dependent fields",
+    "version": "18.0.1.0.0",
     "license": "AGPL-3",
     "maintainer": ["bealdav"],
     "website": "https://github.com/OCA/multi-company",
@@ -10,6 +10,4 @@
     "depends": [
         "base",
     ],
-    "data": [],
-    "installable": True,
 }
