+++ web_field_company_dependent_flag/models/base.py
@@ -9,7 +9,7 @@ class Base(models.AbstractModel):
 
     @api.model
     def _get_view(self, view_id=None, view_type="form", **options):
-        arch, view = super()._get_view(view_id, view_type, **options)
+        arch, view = super()._get_view(view_id=view_id, view_type=view_type, **options)
         if view_type == "form":
             self._update_company_dependent_css(arch)
         return arch, view
